requirements.txt
flask
requests
flask-oauthlib
redis
dnspython
